<< enum typinvalinta ei ole aina selvä juttu, lisäsin static_cast<int> pariin kohtaan varoitusten välttämiseksi.

+++ source/NFmiPressDescription.cpp
@@ -50,7 +50,7 @@ bool NFmiPressDescription::ReadRemaining(void)
 		  *itsLogFile << "*** ERROR: Not valid PressMask: "
 					  << static_cast<char *>(maskName)
 					  << " in space "
-					  << helpEnumSpace 
+					  << static_cast<int>(helpEnumSpace)
 					  << endl;
 		}
